Moved tcp `Connection` struct to new module
